+++ R/EMSS.R
@@ -39,7 +39,7 @@
 #' \item{names_selection}{names of regression coefficients for the selection formula.}
 #' @importFrom mvtnorm rmvnorm
 #' @importFrom sampleSelection selection
-#' @importFrom stats dnorm model.matrix model.response na.pass pnorm qchisq qnorm symnum
+#' @importFrom stats dnorm model.matrix model.response na.pass pnorm qchisq qnorm symnum complete.cases
 #' @examples
 #' data(Smoke, package = "EMSS")
 #' ex1 <- EMSS(response = cigs_intervals ~ educ,
@@ -118,9 +118,17 @@ EMSS <- function(response, selection, data, method="ECM",
 
   x <- model.matrix(mt1, mf1, contrasts.arg = NULL, xlev = NULL)
   w <- model.matrix(mt2, mf2, contrasts.arg = NULL, xlev = NULL)
-  if(any(is.na(x)==TRUE)){
-    message("Warning: There is NA data in the observed characters of the response data.")
+  
+  complete <- complete.cases(x) & complete.cases(w)
+  
+  if(sum(!complete) > 0){
+    x <- x[complete,]
+    w <- w[complete,]
+    y1 <- y1[complete]
+    y2 <- y2[complete]
+    warning("The response data or the selection data is not complete, the missing observations are ignored.")
   }
+  
   x.name <- names(as.data.frame(x))
   w.name <- names(as.data.frame(w))