+++ NEWS.md
@@ -3,3 +3,7 @@
 * We changed the structures of the function EMSS. There are now internal functions for fitting various methods, and also for getting the Q value. 
 * We added a github repository for reporting bugs.
 * Now, users can use tidy type of summary by using arguments in the summary.EMSS function.
+
+# EMSS 1.1.1 - Changes on 12/21/2021
+
+* There is a modification with handling the missing data. We changed the warning message and new approach of handling missing data is adapted to both selection data and response data.